// backend/src/module/common/common.type.ts
export interface Post {
  id: string;
  user_id: string;
  content: string | null;
  image_urls: string[] | null;
  document_id: string | null;
  is_deleted: boolean;
  created_at: Date;
  updated_at: Date;
}

export interface Like {
  id: string;
  post_id: string;
  user_id: string;
  created_at: Date;
}

export interface Comment {
  id: string;
  post_id: string;
  user_id: string;
  content: string;
  created_at: Date;
}